feat: Optimize training and inference with bf16 and increase batch size
This commit introduces several optimizations to the training and inference pipelines and fixes critical bugs.

Key changes:
- Switched training and inference to use `bf16` precision for better performance on compatible hardware. The training script logic was updated to correctly use the `amp_dtype` from the configuration.
- Enabled gradient checkpointing during training to reduce memory usage.
- Increased the training batch size from 2 to 8, made possible by the memory savings from gradient checkpointing.
- Fixed a bug in the inference engine where the precision setting was not being correctly applied. The logic is now simplified and correctly uses the `precision` setting.
- Removed the redundant `use_fp16` flag from the inference configuration to improve clarity.
- Fixed a crash in `train_direct.py` caused by a missing `focal_clip` parameter by adding it to the configuration with a sensible default.
- Corrected a typo from `bfloat16` to `bf16` to match the system's validation rules.

+++ Configuration_System.py
@@ -668,7 +668,7 @@ class InferenceConfig(BaseConfig):
     """Inference configuration"""
     # Model
     model_path: Optional[str] = None
-    precision: str = "fp16"  # Options: "fp32", "fp16", "bf16"
+    precision: str = "bf16"  # Options: "fp32", "fp16", "bf16"
     compile_model: bool = False
     
     # Prediction
